+++ src/pages/SignUpPage/constants/index.ts
@@ -1,0 +1,12 @@
+export const SIGNUP_INITIAL_VALUE = 'init';
+
+export const MSG = {
+  WARNING: {
+    EMPTY: {
+      EMAIL: '이메일을 입력해주세요.',
+    },
+  },
+  ERROR: {
+    DUPLICATE_CHECK: '중복검사를 위해 유저 정보를 가져오는데 실패하였습니다.',
+  },
+};
